+++ tests/SecurePasswordTest.php
@@ -14,10 +14,31 @@ class SecurePasswordTest extends TestCase
         $this->assertIsString($hash);
     }
 
+    public function testCreateHashWithConfig()
+    {
+        $config = [
+            'algo' => HashAlgorithm::ARGON2I,
+            'cost' => 10,
+            'memory_cost' => PASSWORD_ARGON2_DEFAULT_MEMORY_COST,
+            'time_cost' => PASSWORD_ARGON2_DEFAULT_TIME_COST,
+            'threads' => PASSWORD_ARGON2_DEFAULT_THREADS
+        ];
+        
+        $password = new SecurePassword($config);
+        $hash = $password->createHash('my_password')->getHash();
+
+        $this->assertIsString($hash);
+    }
+
     public function testChangePepper()
     {
         $password = new SecurePassword();
-        $password->setPepper('new_pepper');
+        $password->setPepper('new_pepper', 'openssl');
+        $hash = $password->createHash('my_password')->getHash();
+
+        $this->assertIsString($hash);
+
+        $password->setPepper('new_other_pepper', 'sodium');
         $hash = $password->createHash('my_password')->getHash();
 
         $this->assertIsString($hash);
@@ -79,7 +100,12 @@ class SecurePasswordTest extends TestCase
     public function testVerifyHash()
     {
         $hash = '$2y$10$Er0wYRuY7LTYkmWmL8YMMeuxiRIEZ7Vn/8kPb4.aNkzIFRN/N.qG.';
-        $res = (new SecurePassword)->verifyHash('my_password', $hash);
+        $res = (new SecurePassword())->verifyHash('my_password', $hash);
+
+        $this->assertTrue($res);
+
+        $password = new SecurePassword();
+        $res = $password->createHash('my_password')->verifyHash();
 
         $this->assertTrue($res);
     }